requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile -v requirements.in
#
-e file:///Users/dev/Projects/zproc
alabaster==0.7.12         # via sphinx
appnope==0.1.0            # via ipython
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via glom, pytest
babel==2.6.0              # via sphinx
backcall==0.1.0           # via ipython
bleach==3.1.0             # via readme-renderer
boltons==19.1.0           # via face, glom
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0
cloudpickle==1.2.1
decorator==4.3.2          # via ipython, traitlets
docutils==0.14            # via readme-renderer, sphinx
face==19.1.2              # via glom
glom==19.2.0
idna==2.8                 # via requests
imagesize==1.1.0          # via sphinx
ipython-genutils==0.2.0   # via traitlets
ipython==7.3.0
jedi==0.13.3              # via ipython
jinja2==2.10              # via sphinx
markupsafe==1.1.1         # via jinja2
more-itertools==6.0.0     # via pytest
mypy-extensions==0.4.1    # via mypy
mypy==0.670
packaging==19.0           # via sphinx
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.4.0
pkginfo==1.5.0.1          # via twine
pluggy==0.9.0             # via pytest
prompt-toolkit==2.0.9     # via ipython
psutil==5.6.1
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pygments==2.3.1           # via ipython, readme-renderer, sphinx
pyparsing==2.3.1          # via packaging
pytest-repeat==0.8.0
pytest==4.3.0
python-decouple==3.1
pytz==2018.9              # via babel
pyzmq==18.0.1
readme-renderer==24.0     # via twine
requests-toolbelt==0.9.1  # via twine
requests==2.21.0          # via requests-toolbelt, sphinx, twine
six==1.12.0               # via bleach, packaging, pip-tools, prompt-toolkit, pytest, readme-renderer, sphinx, traitlets
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.4.3
sphinx==1.8.5
sphinxcontrib-websupport==1.1.0  # via sphinx
tblib==1.3.2
tqdm==4.31.1              # via twine
traitlets==4.3.2          # via ipython
twine==1.13.0
typed-ast==1.3.1          # via mypy
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via bleach
yappi==1.0
